Implemented gRPC on client and server
Something weird when i send parameters, i do get an image in return but i must have misplaced the params. It takes 5-6 seconds to regenerate an image, might be good enough, ill try making it faster. Streaming the byte array was a lot faster than loading first into memory, can return the image directly through the endpoint this way. If  this works, it can easily be replicated throughout the project