wip grpc WAV chunk streaming
Working on sending wav chunks with gRPC to SignalProcessing microservice for live waveform rendering on the frontend, the server ran but no testing has been done yet. Need to write client code. Maybe it is smart to use WebSockets to stream the spectrogram data to the frontend instead of going throug the C# API, probably the best  solution